Aligned inputs with other modules
It was pointed out in a peer review of this module that parameters were being passed through in a different manor to other modules. Therefore, it was decided to make changes to the Enrichment module so that it would take inputs individually, rather than grouping them into a 'parameters' dictionary.

- Removed parameters dictionary from wrangler and passed through RuntimeVariables individually
- Updated the method to receive these inputs from the event rather than from the parameters dictionary
- Removed parameters dictionary from the tests