Remove redundant install of npm in Docker build
The debsource script in the Dockerfile (deb.nodesource.com/setup_10.x)
installs Nodejs and a compatible version of NPM so one does not need to
install NPM again since that will be installed or upgraded
automatically.

Attempting to do so may try to install npm from another debian
repository which may be incompatible and fail as seen in this
case.

Task: LU-5093